chore: refactored structure
