+++ meowtg/main.ts
@@ -7,8 +7,8 @@ import {getDisplayName} from "telegram/Utils";
 import {LogLevel} from "telegram/extensions/Logger";
 import * as dotenv from "dotenv";
 import SessionManager from "./sessionManager";
-import CommandsProcessor from "./commandsProcessor";
-import ConfigurationWizard from "./configurationWizard";
+import CommandsProcessor from "./command/commandsProcessor";
+import ConfigurationWizard from "./wizard/configurationWizard";
 import PluginsProcessor from "./plugin/pluginsProcessor";
 import PluginsAPI from "./plugin/pluginsApi";
 import {isPrivateMessageNotMine, parseArguments} from "./utils";
